+++ src/main/resources/templates/ad/ad.html
@@ -171,7 +171,7 @@
                 </div>
             </div>
 
-            <div th:if="${#lists.isEmpty(ads)}" style="width: 440px; height: 280px; margin-left: 25%">
+            <!--div th:if="${#lists.isEmpty(ads)}" style="width: 440px; height: 280px; margin-left: 25%">
                 <div class="carousel slide" data-ride="carousel">
                     <div class="carousel-inner">
                         <div class="carousel-item active">
@@ -179,7 +179,7 @@
                         </div>
                     </div>
                 </div>
-            </div>
+            </div-->
         </div>
     </div>
